corrected angle and force

+++ src/apps/project-app.h
@@ -47,8 +47,8 @@ int calc_max_height(double force, double angle);
 void gl_draw_image(unsigned int x, unsigned int y, char first_initial);
 unsigned int gl_plot_image_trajectory(double force, double angle, char first_initial);
 void gl_draw_target(unsigned int leftBound, unsigned int size);
-unsigned int read_force(void);
-unsigned int read_angle(void);
+double read_force(void);
+double read_angle(void);
 
 void angry_nerds_graphics_init(void);
 void angry_nerds_game_init(void);
